requirements.txt
blinker==1.5
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
flask_weasyprint==1.0.0
fonttools==4.38.0
greenlet==2.0.1
html5lib==1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.2
passlib==1.7.4
Pillow==9.4.0
platformdirs==2.6.2
protobuf==3.20.1
pycparser==2.21
pydyf==0.5.0
PyMySQL==1.0.2
pyphen==0.13.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.46
tinycss2==1.2.1
virtualenv==20.17.1
weasyprint==57.2
webencodings==0.5.1
Werkzeug==2.2.2
zopfli==0.2.2
